moves settings to a route
