+++ lib/app_router.gr.dart
@@ -10,85 +10,92 @@
 //
 // ignore_for_file: type=lint
 
-import 'package:auto_route/auto_route.dart' as _i9;
-import 'package:flutter/material.dart' as _i10;
+import 'package:auto_route/auto_route.dart' as _i10;
+import 'package:flutter/material.dart' as _i11;
 
 import 'screens/authentication/authentication_screen.dart' as _i2;
-import 'screens/feedback/feedback_screen.dart' as _i5;
-import 'screens/meal/meal_screen.dart' as _i7;
-import 'screens/meal_create/meal_create_screen.dart' as _i6;
-import 'screens/meal_select/meal_select.dart' as _i3;
-import 'screens/onboarding/onboarding_screen.dart' as _i4;
+import 'screens/feedback/feedback_screen.dart' as _i6;
+import 'screens/meal/meal_screen.dart' as _i8;
+import 'screens/meal_create/meal_create_screen.dart' as _i7;
+import 'screens/meal_select/meal_select.dart' as _i4;
+import 'screens/onboarding/onboarding_screen.dart' as _i5;
+import 'screens/settings/settings_screen.dart' as _i3;
 import 'screens/tab_navigation/home_screen.dart' as _i1;
-import 'screens/unknown_route/unknown_route_screen.dart' as _i8;
+import 'screens/unknown_route/unknown_route_screen.dart' as _i9;
 
-class AppRouter extends _i9.RootStackRouter {
-  AppRouter([_i10.GlobalKey<_i10.NavigatorState>? navigatorKey])
+class AppRouter extends _i10.RootStackRouter {
+  AppRouter([_i11.GlobalKey<_i11.NavigatorState>? navigatorKey])
       : super(navigatorKey);
 
   @override
-  final Map<String, _i9.PageFactory> pagesMap = {
+  final Map<String, _i10.PageFactory> pagesMap = {
     HomeScreenRoute.name: (routeData) {
-      return _i9.CupertinoPageX<dynamic>(
+      return _i10.CupertinoPageX<dynamic>(
           routeData: routeData, child: const _i1.HomeScreen());
     },
     AuthenticationScreenRoute.name: (routeData) {
-      return _i9.CupertinoPageX<dynamic>(
+      return _i10.CupertinoPageX<dynamic>(
           routeData: routeData, child: const _i2.AuthenticationScreen());
+    },
+    SettingsScreenRoute.name: (routeData) {
+      return _i10.CupertinoPageX<dynamic>(
+          routeData: routeData, child: const _i3.SettingsScreen());
     },
     MealSelectScreenRoute.name: (routeData) {
       final args = routeData.argsAs<MealSelectScreenRouteArgs>();
-      return _i9.CupertinoPageX<dynamic>(
+      return _i10.CupertinoPageX<dynamic>(
           routeData: routeData,
-          child: _i3.MealSelectScreen(
+          child: _i4.MealSelectScreen(
               date: args.date, isLunch: args.isLunch, key: args.key));
     },
     OnboardingScreenRoute.name: (routeData) {
       final args = routeData.argsAs<OnboardingScreenRouteArgs>(
           orElse: () => const OnboardingScreenRouteArgs());
-      return _i9.CupertinoPageX<dynamic>(
-          routeData: routeData, child: _i4.OnboardingScreen(key: args.key));
+      return _i10.CupertinoPageX<dynamic>(
+          routeData: routeData, child: _i5.OnboardingScreen(key: args.key));
     },
     FeedbackScreenRoute.name: (routeData) {
-      return _i9.CupertinoPageX<dynamic>(
-          routeData: routeData, child: const _i5.FeedbackScreen());
+      return _i10.CupertinoPageX<dynamic>(
+          routeData: routeData, child: const _i6.FeedbackScreen());
     },
     MealCreateScreenRoute.name: (routeData) {
       final args = routeData.argsAs<MealCreateScreenRouteArgs>();
-      return _i9.CupertinoPageX<dynamic>(
+      return _i10.CupertinoPageX<dynamic>(
           routeData: routeData,
-          child: _i6.MealCreateScreen(id: args.id, key: args.key));
+          child: _i7.MealCreateScreen(id: args.id, key: args.key));
     },
     MealScreenRoute.name: (routeData) {
       final args = routeData.argsAs<MealScreenRouteArgs>();
-      return _i9.CupertinoPageX<dynamic>(
+      return _i10.CupertinoPageX<dynamic>(
           routeData: routeData,
-          child: _i7.MealScreen(id: args.id, key: args.key));
+          child: _i8.MealScreen(id: args.id, key: args.key));
     },
     UnknownRouteScreenRoute.name: (routeData) {
-      return _i9.CupertinoPageX<dynamic>(
-          routeData: routeData, child: const _i8.UnknownRouteScreen());
+      return _i10.CupertinoPageX<dynamic>(
+          routeData: routeData, child: const _i9.UnknownRouteScreen());
     }
   };
 
   @override
-  List<_i9.RouteConfig> get routes => [
-        _i9.RouteConfig(HomeScreenRoute.name, path: '/'),
-        _i9.RouteConfig(AuthenticationScreenRoute.name,
+  List<_i10.RouteConfig> get routes => [
+        _i10.RouteConfig(HomeScreenRoute.name, path: '/'),
+        _i10.RouteConfig(AuthenticationScreenRoute.name,
             path: '/authentication-screen'),
-        _i9.RouteConfig(MealSelectScreenRoute.name,
+        _i10.RouteConfig(SettingsScreenRoute.name, path: '/settings-screen'),
+        _i10.RouteConfig(MealSelectScreenRoute.name,
             path: '/meal-select-screen'),
-        _i9.RouteConfig(OnboardingScreenRoute.name, path: '/onboarding-screen'),
-        _i9.RouteConfig(FeedbackScreenRoute.name, path: '/feedback-screen'),
-        _i9.RouteConfig(MealCreateScreenRoute.name, path: '/meal-create/:id'),
-        _i9.RouteConfig(MealScreenRoute.name, path: '/meal/:id'),
-        _i9.RouteConfig(UnknownRouteScreenRoute.name, path: '*')
+        _i10.RouteConfig(OnboardingScreenRoute.name,
+            path: '/onboarding-screen'),
+        _i10.RouteConfig(FeedbackScreenRoute.name, path: '/feedback-screen'),
+        _i10.RouteConfig(MealCreateScreenRoute.name, path: '/meal-create/:id'),
+        _i10.RouteConfig(MealScreenRoute.name, path: '/meal/:id'),
+        _i10.RouteConfig(UnknownRouteScreenRoute.name, path: '*')
       ];
 }
 
 /// generated route for
 /// [_i1.HomeScreen]
-class HomeScreenRoute extends _i9.PageRouteInfo<void> {
+class HomeScreenRoute extends _i10.PageRouteInfo<void> {
   const HomeScreenRoute() : super(HomeScreenRoute.name, path: '/');
 
   static const String name = 'HomeScreenRoute';
@@ -96,7 +103,7 @@ class HomeScreenRoute extends _i9.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i2.AuthenticationScreen]
-class AuthenticationScreenRoute extends _i9.PageRouteInfo<void> {
+class AuthenticationScreenRoute extends _i10.PageRouteInfo<void> {
   const AuthenticationScreenRoute()
       : super(AuthenticationScreenRoute.name, path: '/authentication-screen');
 
@@ -104,11 +111,20 @@ class AuthenticationScreenRoute extends _i9.PageRouteInfo<void> {
 }
 
 /// generated route for
-/// [_i3.MealSelectScreen]
+/// [_i3.SettingsScreen]
+class SettingsScreenRoute extends _i10.PageRouteInfo<void> {
+  const SettingsScreenRoute()
+      : super(SettingsScreenRoute.name, path: '/settings-screen');
+
+  static const String name = 'SettingsScreenRoute';
+}
+
+/// generated route for
+/// [_i4.MealSelectScreen]
 class MealSelectScreenRoute
-    extends _i9.PageRouteInfo<MealSelectScreenRouteArgs> {
+    extends _i10.PageRouteInfo<MealSelectScreenRouteArgs> {
   MealSelectScreenRoute(
-      {required DateTime date, required bool isLunch, _i10.Key? key})
+      {required DateTime date, required bool isLunch, _i11.Key? key})
       : super(MealSelectScreenRoute.name,
             path: '/meal-select-screen',
             args: MealSelectScreenRouteArgs(
@@ -125,7 +141,7 @@ class MealSelectScreenRouteArgs {
 
   final bool isLunch;
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   @override
   String toString() {
@@ -134,10 +150,10 @@ class MealSelectScreenRouteArgs {
 }
 
 /// generated route for
-/// [_i4.OnboardingScreen]
+/// [_i5.OnboardingScreen]
 class OnboardingScreenRoute
-    extends _i9.PageRouteInfo<OnboardingScreenRouteArgs> {
-  OnboardingScreenRoute({_i10.Key? key})
+    extends _i10.PageRouteInfo<OnboardingScreenRouteArgs> {
+  OnboardingScreenRoute({_i11.Key? key})
       : super(OnboardingScreenRoute.name,
             path: '/onboarding-screen',
             args: OnboardingScreenRouteArgs(key: key));
@@ -148,7 +164,7 @@ class OnboardingScreenRoute
 class OnboardingScreenRouteArgs {
   const OnboardingScreenRouteArgs({this.key});
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   @override
   String toString() {
@@ -157,8 +173,8 @@ class OnboardingScreenRouteArgs {
 }
 
 /// generated route for
-/// [_i5.FeedbackScreen]
-class FeedbackScreenRoute extends _i9.PageRouteInfo<void> {
+/// [_i6.FeedbackScreen]
+class FeedbackScreenRoute extends _i10.PageRouteInfo<void> {
   const FeedbackScreenRoute()
       : super(FeedbackScreenRoute.name, path: '/feedback-screen');
 
@@ -166,10 +182,10 @@ class FeedbackScreenRoute extends _i9.PageRouteInfo<void> {
 }
 
 /// generated route for
-/// [_i6.MealCreateScreen]
+/// [_i7.MealCreateScreen]
 class MealCreateScreenRoute
-    extends _i9.PageRouteInfo<MealCreateScreenRouteArgs> {
-  MealCreateScreenRoute({required String id, _i10.Key? key})
+    extends _i10.PageRouteInfo<MealCreateScreenRouteArgs> {
+  MealCreateScreenRoute({required String id, _i11.Key? key})
       : super(MealCreateScreenRoute.name,
             path: '/meal-create/:id',
             args: MealCreateScreenRouteArgs(id: id, key: key));
@@ -182,7 +198,7 @@ class MealCreateScreenRouteArgs {
 
   final String id;
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   @override
   String toString() {
@@ -191,9 +207,9 @@ class MealCreateScreenRouteArgs {
 }
 
 /// generated route for
-/// [_i7.MealScreen]
-class MealScreenRoute extends _i9.PageRouteInfo<MealScreenRouteArgs> {
-  MealScreenRoute({required String id, _i10.Key? key})
+/// [_i8.MealScreen]
+class MealScreenRoute extends _i10.PageRouteInfo<MealScreenRouteArgs> {
+  MealScreenRoute({required String id, _i11.Key? key})
       : super(MealScreenRoute.name,
             path: '/meal/:id', args: MealScreenRouteArgs(id: id, key: key));
 
@@ -205,7 +221,7 @@ class MealScreenRouteArgs {
 
   final String id;
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   @override
   String toString() {
@@ -214,8 +230,8 @@ class MealScreenRouteArgs {
 }
 
 /// generated route for
-/// [_i8.UnknownRouteScreen]
-class UnknownRouteScreenRoute extends _i9.PageRouteInfo<void> {
+/// [_i9.UnknownRouteScreen]
+class UnknownRouteScreenRoute extends _i10.PageRouteInfo<void> {
   const UnknownRouteScreenRoute()
       : super(UnknownRouteScreenRoute.name, path: '*');
